factored out fetches
